durable-func/requirements.txt
aenum
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
azure-ai-textanalytics
azure-common
azure-core
azure-functions
azure-functions-durable
azure-storage-blob
certifi
cffi
charset-normalizer
cryptography
Deprecated
distro
exceptiongroup
frozenlist
furl
future
h11
httpcore
httpx
idna
isodate
line-bot-sdk
multidict
openai==1.2.3
orderedmultidict
pycparser
pydantic
pydantic_core
python-dateutil
requests
six
sniffio
tqdm
typing_extensions
urllib3
wrapt
yarl
